refactor: extract CosyLogo + useLanguageChange, add lang toggle to landing page
- New CosyLogo component (optional link) replaces duplicated logo markup
  in AppHeader and LandingNav
- New useLanguageChange hook replaces duplicated language logic
  in AppHeader and AuthPageLayout
- LandingNav: logo is no longer clickable, LanguageMenu added
- AppHeader and AuthPageLayout now use the shared hook
